Move API related code to a sub-module + update tests
