+++ custom_components/irm_kmi/__init__.py
@@ -9,9 +9,9 @@ from homeassistant.exceptions import ConfigEntryError
 
 from .const import (CONF_DARK_MODE, CONF_LANGUAGE_OVERRIDE, CONF_STYLE,
                     CONF_USE_DEPRECATED_FORECAST, CONFIG_FLOW_VERSION, DOMAIN,
-                    OPTION_DEPRECATED_FORECAST_NOT_USED, OPTION_STYLE_STD,
-                    PLATFORMS)
+                    OPTION_DEPRECATED_FORECAST_NOT_USED, PLATFORMS)
 from .coordinator import IrmKmiCoordinator
+from .irm_kmi_api.const import OPTION_STYLE_STD
 from .weather import IrmKmiWeather
 
 _LOGGER = logging.getLogger(__name__)
