+++ tests/test_coordinator.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timedelta
+from zoneinfo import ZoneInfo
 
 from freezegun import freeze_time
 from homeassistant.components.weather import (ATTR_CONDITION_CLOUDY,
@@ -9,11 +10,11 @@ from pytest_homeassistant_custom_component.common import MockConfigEntry
 
 from custom_components.irm_kmi.const import CONF_LANGUAGE_OVERRIDE
 from custom_components.irm_kmi.coordinator import IrmKmiCoordinator
-from custom_components.irm_kmi.data import (CurrentWeatherData, IrmKmiForecast,
-                                            ProcessedCoordinatorData)
-from custom_components.irm_kmi.pollen import PollenParser
-from custom_components.irm_kmi.radar_data import IrmKmiRadarForecast
-from tests.conftest import get_api_data
+from custom_components.irm_kmi.irm_kmi_api.api import IrmKmiApiClientHa
+from custom_components.irm_kmi.irm_kmi_api.data import (CurrentWeatherData, IrmKmiForecast, IrmKmiRadarForecast)
+from custom_components.irm_kmi.data import ProcessedCoordinatorData
+from custom_components.irm_kmi.irm_kmi_api.pollen import PollenParser
+from tests.conftest import get_api_data, get_api_with_data
 
 
 async def test_jules_forgot_to_revert_update_interval_before_pushing(
@@ -27,19 +28,16 @@ async def test_jules_forgot_to_revert_update_interval_before_pushing(
 
 @freeze_time(datetime.fromisoformat('2024-01-12T07:10:00+00:00'))
 async def test_warning_data(
-        hass: HomeAssistant,
         mock_config_entry: MockConfigEntry
 ) -> None:
-    api_data = get_api_data("be_forecast_warning.json")
-    coordinator = IrmKmiCoordinator(hass, mock_config_entry)
+    api = get_api_with_data("be_forecast_warning.json")
 
-    result = coordinator.warnings_from_data(api_data.get('for', {}).get('warning'))
+    result = api.get_warnings(lang='en')
 
     assert isinstance(result, list)
     assert len(result) == 2
 
     first = result[0]
-
     assert first.get('starts_at').replace(tzinfo=None) < datetime.now()
     assert first.get('ends_at').replace(tzinfo=None) > datetime.now()
 
@@ -51,8 +49,9 @@ async def test_warning_data(
 
 @freeze_time(datetime.fromisoformat('2023-12-26T17:30:00+00:00'))
 async def test_current_weather_be() -> None:
-    api_data = get_api_data("forecast.json")
-    result = await IrmKmiCoordinator.current_weather_from_data(api_data)
+    api = get_api_with_data("forecast.json")
+    tz = ZoneInfo("Europe/Brussels")
+    result = await api.get_current_weather(tz)
 
     expected = CurrentWeatherData(
         condition=ATTR_CONDITION_CLOUDY,
@@ -69,8 +68,9 @@ async def test_current_weather_be() -> None:
 
 @freeze_time(datetime.fromisoformat("2023-12-28T15:30:00"))
 async def test_current_weather_nl() -> None:
-    api_data = get_api_data("forecast_nl.json")
-    result = await IrmKmiCoordinator.current_weather_from_data(api_data)
+    api = get_api_with_data("forecast_nl.json")
+    tz = ZoneInfo("Europe/Brussels")
+    result = await api.get_current_weather(tz)
 
     expected = CurrentWeatherData(
         condition=ATTR_CONDITION_CLOUDY,
@@ -90,11 +90,10 @@ async def test_daily_forecast(
         hass: HomeAssistant,
         mock_config_entry: MockConfigEntry
 ) -> None:
-    api_data = get_api_data("forecast.json").get('for', {}).get('daily')
-    await hass.config_entries.async_add(mock_config_entry)
-    hass.config_entries.async_update_entry(mock_config_entry, data=mock_config_entry.data | {CONF_LANGUAGE_OVERRIDE: 'fr'})
-    coordinator = IrmKmiCoordinator(hass, mock_config_entry)
-    result = await coordinator.daily_list_to_forecast(api_data)
+    api = get_api_with_data("forecast.json")
+    tz = ZoneInfo("Europe/Brussels")
+
+    result = await api.get_daily_forecast(tz, 'fr')
 
     assert isinstance(result, list)
     assert len(result) == 8
@@ -121,8 +120,9 @@ async def test_daily_forecast(
 
 @freeze_time(datetime.fromisoformat('2023-12-26T18:30:00+01:00'))
 async def test_hourly_forecast() -> None:
-    api_data = get_api_data("forecast.json").get('for', {}).get('hourly')
-    result = await IrmKmiCoordinator.hourly_list_to_forecast(api_data)
+    api = get_api_with_data("forecast.json")
+    tz = ZoneInfo("Europe/Brussels")
+    result = await api.get_hourly_forecast(tz)
 
     assert isinstance(result, list)
     assert len(result) == 49
@@ -146,8 +146,10 @@ async def test_hourly_forecast() -> None:
 
 @freeze_time(datetime.fromisoformat('2024-05-31T01:50:00+02:00'))
 async def test_hourly_forecast_bis() -> None:
-    api_data = get_api_data("no-midnight-bug-31-05-2024T01-55.json").get('for', {}).get('hourly')
-    result = await IrmKmiCoordinator.hourly_list_to_forecast(api_data)
+    api = get_api_with_data("no-midnight-bug-31-05-2024T01-55.json")
+    tz = ZoneInfo("Europe/Brussels")
+
+    result = await api.get_hourly_forecast(tz)
 
     assert isinstance(result, list)
 
@@ -163,8 +165,10 @@ async def test_hourly_forecast_bis() -> None:
 @freeze_time(datetime.fromisoformat('2024-05-31T00:10:00+02:00'))
 async def test_hourly_forecast_midnight_bug() -> None:
     # Related to https://github.com/jdejaegh/irm-kmi-ha/issues/38
-    api_data = get_api_data("midnight-bug-31-05-2024T00-13.json").get('for', {}).get('hourly')
-    result = await IrmKmiCoordinator.hourly_list_to_forecast(api_data)
+    api = get_api_with_data("midnight-bug-31-05-2024T00-13.json")
+    tz = ZoneInfo("Europe/Brussels")
+
+    result = await api.get_hourly_forecast(tz)
 
     assert isinstance(result, list)
 
@@ -200,10 +204,10 @@ async def test_daily_forecast_midnight_bug(
         hass: HomeAssistant,
         mock_config_entry: MockConfigEntry
 ) -> None:
-    coordinator = IrmKmiCoordinator(hass, mock_config_entry)
+    api = get_api_with_data("midnight-bug-31-05-2024T00-13.json")
+    tz = ZoneInfo("Europe/Brussels")
 
-    api_data = get_api_data("midnight-bug-31-05-2024T00-13.json").get('for', {}).get('daily')
-    result = await coordinator.daily_list_to_forecast(api_data)
+    result = await api.get_daily_forecast(tz, 'en')
 
     assert result[0]['datetime'] == '2024-05-31'
     assert not result[0]['is_daytime']
@@ -221,19 +225,11 @@ async def test_daily_forecast_midnight_bug(
 async def test_refresh_succeed_even_when_pollen_and_radar_fail(
         hass: HomeAssistant,
         mock_config_entry: MockConfigEntry,
-        mock_irm_kmi_api_works_but_pollen_and_radar_fail
 ):
-    hass.states.async_set(
-        "zone.home",
-        0,
-        {"latitude": 50.738681639, "longitude": 4.054077148},
-    )
-    hass.config.config_dir = "."
-    mock_config_entry.add_to_hass(hass)
-
     coordinator = IrmKmiCoordinator(hass, mock_config_entry)
+    coordinator._api._api_data = get_api_data("forecast.json")
 
-    result = await coordinator._async_update_data()
+    result = await coordinator.process_api_data()
 
     assert result.get('current_weather').get('condition') == ATTR_CONDITION_CLOUDY
 
@@ -250,7 +246,7 @@ async def test_refresh_succeed_even_when_pollen_and_radar_fail(
         pollen={'foo': 'bar'}
     )
     coordinator.data = existing_data
-    result = await coordinator._async_update_data()
+    result = await coordinator.process_api_data()
 
     assert result.get('current_weather').get('condition') == ATTR_CONDITION_CLOUDY
 
@@ -260,8 +256,8 @@ async def test_refresh_succeed_even_when_pollen_and_radar_fail(
 
 
 def test_radar_forecast() -> None:
-    api_data = get_api_data("forecast.json")
-    result = IrmKmiCoordinator.radar_list_to_forecast(api_data.get('animation'))
+    api = get_api_with_data("forecast.json")
+    result = api.get_radar_forecast()
 
     expected = [
         IrmKmiRadarForecast(datetime="2023-12-26T17:00:00+01:00", native_precipitation=0, might_rain=False,
@@ -292,8 +288,8 @@ def test_radar_forecast() -> None:
 
 
 def test_radar_forecast_rain_interval() -> None:
-    api_data = get_api_data('forecast_with_rain_on_radar.json')
-    result = IrmKmiCoordinator.radar_list_to_forecast(api_data.get('animation'))
+    api = get_api_with_data('forecast_with_rain_on_radar.json')
+    result = api.get_radar_forecast()
 
     _12 = IrmKmiRadarForecast(
         datetime='2024-05-30T18:00:00+02:00',
@@ -322,10 +318,10 @@ async def test_datetime_daily_forecast_nl(
         hass: HomeAssistant,
         mock_config_entry: MockConfigEntry
 ) -> None:
-    api_data = get_api_data("forecast_ams_no_ww.json").get('for', {}).get('daily')
+    api = get_api_with_data("forecast_ams_no_ww.json")
+    tz = ZoneInfo("Europe/Brussels")
 
-    coordinator = IrmKmiCoordinator(hass, mock_config_entry)
-    result = await coordinator.daily_list_to_forecast(api_data)
+    result = await api.get_daily_forecast(tz, 'en')
 
     assert result[0]['datetime'] == '2024-06-09'
     assert result[0]['is_daytime']
@@ -339,8 +335,10 @@ async def test_datetime_daily_forecast_nl(
 
 @freeze_time("2024-06-09T13:40:00+00:00")
 async def test_current_condition_forecast_nl() -> None:
-    api_data = get_api_data("forecast_ams_no_ww.json")
-    result = await IrmKmiCoordinator.current_weather_from_data(api_data)
+    api = get_api_with_data("forecast_ams_no_ww.json")
+    tz = ZoneInfo("Europe/Brussels")
+
+    result = await api.get_current_weather(tz)
 
     expected = CurrentWeatherData(
         condition=ATTR_CONDITION_PARTLYCLOUDY,
@@ -359,10 +357,10 @@ async def test_sunrise_sunset_nl(
         hass: HomeAssistant,
         mock_config_entry: MockConfigEntry
 ) -> None:
-    api_data = get_api_data("forecast_ams_no_ww.json").get('for', {}).get('daily')
+    api = get_api_with_data("forecast_ams_no_ww.json")
+    tz = ZoneInfo("Europe/Brussels")
 
-    coordinator = IrmKmiCoordinator(hass, mock_config_entry)
-    result = await coordinator.daily_list_to_forecast(api_data)
+    result = await api.get_daily_forecast(tz, 'en')
 
     assert result[0]['sunrise'] == '2024-06-09T05:19:28+02:00'
     assert result[0]['sunset'] == '2024-06-09T22:01:09+02:00'
@@ -379,10 +377,10 @@ async def test_sunrise_sunset_be(
         hass: HomeAssistant,
         mock_config_entry: MockConfigEntry
 ) -> None:
-    api_data = get_api_data("forecast.json").get('for', {}).get('daily')
+    api = get_api_with_data("forecast.json")
+    tz = ZoneInfo("Europe/Brussels")
 
-    coordinator = IrmKmiCoordinator(hass, mock_config_entry)
-    result = await coordinator.daily_list_to_forecast(api_data)
+    result = await api.get_daily_forecast(tz, 'en')
 
     assert result[1]['sunrise'] == '2023-12-27T08:44:00+01:00'
     assert result[1]['sunset'] == '2023-12-27T16:43:00+01:00'
